static_cast update

+++ src/transfers/transfers_ltu.cpp
@@ -294,7 +294,7 @@ double
 trame::transfers::ltu::UW(const double Ws, const int x_ind, const int y_ind)
 const
 {
-    return - Psi((double) 0.0,-Ws,x_ind,y_ind);
+    return - Psi(0.0,-Ws,x_ind,y_ind);
 }
 
 arma::mat 
@@ -318,7 +318,7 @@ double
 trame::transfers::ltu::VW(const double Ws, const int x_ind, const int y_ind)
 const
 {
-    return - Psi(Ws,(double) 0.0,x_ind,y_ind);
+    return - Psi(Ws,0.0,x_ind,y_ind);
 }
 
 // dw
